requirements.txt
opencv-python
Pillow
Flask
matplotlib